requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
convertapi==1.4.0
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
idna==3.3
pyaes==1.6.1
pyasn1==0.4.8
python-decouple==3.5
requests==2.27.1
rsa==4.8
Telethon==1.24.0
urllib3==1.26.8
Werkzeug==2.0.3
